api/requirements.txt
fastapi==0.103.1
rdkit-pypi==2022.9.5
numpy==1.25.2
pydantic==2.3.0
python-multipart==0.0.6
uvicorn==0.23.2